Add dist analysis to "Automated Loan Prediction.ipynb"
Now that we already explored the characteristics of the loan dataset we have,
I now study the distribution of the variables ApplicantIncome and LoanAmount.

Using histograms and box plots, we find out that both ApplicantIncome and
LoanAmount have extreme values, which might be explained by the income
disparity in the society, probably driven by the difference in education
levels in the dataset.